Calculate hierarchy statistics

Calculates various statistics about the code hierarchy including the total number of themes and codes, maximum depth, and distribution of codes across themes.

Calculates various statistics about the code hierarchy including the total number of themes and codes, maximum depth, and distribution of codes across themes.

Usage

calculate_hierarchy_stats(node)

calculate_hierarchy_stats(node)

Arguments

node

Root node of the hierarchy tree

Value

A list containing hierarchy statistics:

total_themes: Total number of themes in the hierarchy

total_codes: Total number of codes in the hierarchy

max_depth: Maximum depth of the hierarchy tree

codes_per_theme: List showing number of codes in each theme

average_codes_per_theme: Average number of codes per theme

A list containing hierarchy statistics:

total_themes: Total number of themes in the hierarchy

total_codes: Total number of codes in the hierarchy

max_depth: Maximum depth of the hierarchy tree

codes_per_theme: List showing number of codes in each theme

average_codes_per_theme: Average number of codes per theme

Examples

if (FALSE) { # \dontrun{
# Create sample hierarchy
root <- Node$new("Root")
theme1 <- root$AddChild("Theme1")
theme1$AddChild("Code1")
theme1$AddChild("Code2")

# Calculate statistics
stats <- calculate_hierarchy_stats(root)
print(stats$total_codes)  # Should print 2
} # }

if (FALSE) { # \dontrun{
# Create sample hierarchy
root <- Node$new("Root")
theme1 <- root$AddChild("Theme1")
theme1$AddChild("Code1")
theme1$AddChild("Code2")

# Calculate statistics
stats <- calculate_hierarchy_stats(root)
print(stats$total_codes)  # Should print 2
} # }