Get Code Paths in Tree

Recursively extracts all full paths from a hierarchical code tree structure. Each path represents the complete hierarchy from root to the given node, with levels separated by forward slashes.

Usage

get_code_paths(node)

Arguments

node

Node in the code tree. Must be a data.tree Node object with the following structure:

name: Character string representing the code name

path: Vector of node names from root to current node

children: List of child nodes (if any)

Value

Character vector of full paths for all codes, excluding the root node

Examples

if (FALSE) { # \dontrun{
# Create a sample code tree using data.tree
library(data.tree)
root <- Node$new("Categories")
methods <- root$AddChild("Methods")
quant <- methods$AddChild("Quantitative")
qual <- methods$AddChild("Qualitative")
quant$AddChild("Survey")
qual$AddChild("Interview")

# Get all code paths
paths <- get_code_paths(root)
print(paths)
# Returns:
# c("Categories/Methods",
#   "Categories/Methods/Quantitative",
#   "Categories/Methods/Quantitative/Survey",
#   "Categories/Methods/Qualitative",
#   "Categories/Methods/Qualitative/Interview")
} # }